Errors about duplicate status are just warnings now
With this if the cache is corrupt or something just remove the file and
rerun `update_twitter.py`.  it will try to post them but complain that
the same post was created already and throw a warning, but it will add
it to cache.

Also fix a couple documentation bugs